Changed parameters for the hpc_performance_test suite to be more aligned with a sanity check (i.e. take 15 minutes versus 8 hours to run).  This entailed removing MATRIX_SIZES.conf.PROD, downsizing the COUNT and SIZE of the hashes generated by fibonacci_hashtest.py, and lowering the maximum exection time of some tests from 8 to 2 hours.
Created a standard submission script for Grid Engine clusters.

Fixed a bug with the src alias not appending properly to
$EC2_HOME/.bashrc.

Modified the jumphost IAM role and policy creation to enable the
jumphost to create and delete cluster stack entities in multiple
regions.

Updated the performance documentation with updated Github markup for
clarity and readability.

Updated README.md and docs/README to reflect the changes in this commit.