Added "--debug_mode" functionality to:
ParallelClusterMaker/ClusterMaker/make-pcluster.py

Removed "True" and "False" as available options for boolean cluster
parameters in both ClusterMaker and JumphostMaker.  Boolean values can now
only be "true" or "false" to make typing command line options easier for
HPC operators.  This is not really using Python booleans properly so at
some point, we will consider converting to "--enable-feature" and
"--disable-feature" to make it more obvious (and Pythonic) when something
is True or False.