Removed references to "centos7" in the access_cluster.py script.  This will be fixed, if needed, at a later date in a future release.
Defined a dictionary of default values for the master and compute instances
to make it easier for operators to just let AWS Batch handle optimization of
the run environment.

Fixed a bug with ansible_python_intrepeter that was preventing
simultaneous usage of this toolkit on both EC2 jumphosts and OSX.

Extended the basic hpc_performance tests to support Slurm.

Updated the documentation to reflect these changes.  Note that the
performance docs have *not* been updated yet to include Slurm support.