Updated the .gitignore file for more specificity with the files and directories in this repo that are not to be checked in.
Added support for tagging cluster resources with "project_id."

Changed the names of the aux_data scripts to be more descriptive.

Extended support for AWS Batch and spot instances.

Increased Ansible verbosity when debug_mode is enabled.

Made some cosmetic changes to the bottom half of the make-cluster and
delete-cluster scripts to improve readability and the overall
look-and-feel of the code.